feat: spec boost Conduit and WardingWrath and ArmOfTheAlmighty and VibrantOrbs fix: army bone on spawn

+++ src/main/java/com/ebicep/warlords/abilities/AvengersWrath.java
@@ -34,7 +34,7 @@ public class AvengersWrath extends AbstractAbility implements OrangeAbilityIcon,
     private int tickDuration = 240;
     private float energyPerSecond = 20;
     private int maxTargets = 2;
-    private int hitRadius = 5;
+    private float hitRadius = 5;
 
     public AvengersWrath() {
         super(AbstractAbilityBuilder.create("avengersWrath").pvp());
@@ -160,11 +160,11 @@ public class AvengersWrath extends AbstractAbility implements OrangeAbilityIcon,
         this.maxTargets = maxTargets;
     }
 
-    public int getHitRadius() {
+    public float getHitRadius() {
         return hitRadius;
     }
 
-    public void setHitRadius(int hitRadius) {
+    public void setHitRadius(float hitRadius) {
         this.hitRadius = hitRadius;
     }
